refactored python static functions and constants out to separate files
